refact(Exception처리 변경) : @ControllerAdvice를 이용해 Exception한 파일에서 처리
CustomException 만들고 GlobalExceptionHandler.java라는 파일에서 처리

CommentSaveException.java=====
@ControllerAdvice애노테이션 제거
간단하게 CustomException을 만드는 방향으로 변경

CommentService.java=====
CommentSaveException에 에러 메세지 추가

GlobalExceptionHandler.java=====
GlobalExceptionHandler.java에서 customException을 잡아서 처리

################
# <타입> (scope): <제목> 의 형식으로 제목을 아래 공백줄에 작성
# 꼬릿말(footer)을 아랫줄에 작성 (현재 커밋과 관련된 이슈 번호 추가 등)
# 예) Close #7
# feat : 새로운 기능 추가
# fix : 버그 수정
# docs : 문서 수정
# test : 테스트 코드 추가
# refact : 코드 리팩토링
# style : 코드 의미에 영향을 주지 않는 변경사항
# chore : 빌드 부분 혹은 패키지 매니저 수정사항

################
# <타입> : <제목> 의 형식으로 제목을 아래 공백줄에 작성
# 제목은 50자 이내 / 변경사항이 "무엇"인지 명확히 작성 / 끝에 마침표 금지
# 예) feat : 로그인 기능 추가

# 바로 아래 공백은 지우지 마세요 (제목과 본문의 분리를 위함)

################
# 본문(구체적인 내용)을 아랫줄에 작성
# 여러 줄의 메시지를 작성할 땐 "-"로 구분 (한 줄은 72자 이내)

################
# feat : 새로운 기능 추가
# fix : 버그 수정
# docs : 문서 수정
# test : 테스트 코드 추가
# refact : 코드 리팩토링
# style : 코드 의미에 영향을 주지 않는 변경사항
# chore : 빌드 부분 혹은 패키지 매니저 수정사항
################
